+++ public/components/agents/vuls/inventory/detail.tsx
@@ -32,7 +32,7 @@ import { getIndexPattern } from '../../../overview/mitre/lib';
 import moment from 'moment-timezone';
 import { AppNavigate } from '../../../../react-services/app-navigate';
 import { TruncateHorizontalComponents } from '../../../common/util';
-import { getDataPlugin,getUiSettings } from '../../../../kibana-services';
+import { getDataPlugin, getUiSettings } from '../../../../kibana-services';
 import { FilterManager } from '../../../../../../../src/plugins/data/public/';
 
 export class Details extends Component {
@@ -112,7 +112,7 @@ export class Details extends Component {
         name: 'Architecture',
         icon: 'node',
         link: false,
-      }
+      },
     ];
   }
 
@@ -138,7 +138,11 @@ export class Details extends Component {
     const { cve } = this.props.currentItem;
     const filters = [
       {
-        ...buildPhraseFilter({ name: 'data.vulnerability.cve', type: 'text' }, cve, this.indexPattern),
+        ...buildPhraseFilter(
+          { name: 'data.vulnerability.cve', type: 'text' },
+          cve,
+          this.indexPattern
+        ),
         $state: { store: 'appState' },
       },
     ];
@@ -151,7 +155,6 @@ export class Details extends Component {
     const { filterManager } = getDataPlugin().query;
     const _filters = filterManager.getFilters();
     if (_filters && _filters.length) {
-     
       filterManager.addFilters([filters]);
       const scope = await ModulesHelper.getDiscoverScope();
       scope.updateQueryAndFetch && scope.updateQueryAndFetch({ query: null });
@@ -188,7 +191,7 @@ export class Details extends Component {
       }
       var link = (item.link && !['events', 'extern'].includes(view)) || false;
       const agentPlatform = ((this.props.agent || {}).os || {}).platform;
-      
+
       if (!item.onlyLinux || (item.onlyLinux && this.props.agent && agentPlatform !== 'windows')) {
         let className = item.checksum ? 'detail-value detail-value-checksum' : 'detail-value';
         className += item.field === 'perm' ? ' detail-value-perm' : '';
@@ -303,7 +306,6 @@ export class Details extends Component {
     return value;
   }
 
-
   render() {
     const { type, implicitFilters, view, currentItem, agent } = this.props;
     const id = `${currentItem.name}-${currentItem.cve}-${currentItem.architecture}-${currentItem.version}`;
@@ -359,11 +361,20 @@ export class Details extends Component {
                 kbnSearchBar
                 shareFilterManager={this.discoverFilterManager}
                 initialColumns={[
-                  'icon',
-                  'timestamp',
-                  'rule.description',
-                  'rule.level',
-                  'rule.id',
+                  { field: 'icon' },
+                  { field: 'timestamp' },
+                  { field: 'agent.id', label: 'Agent' },
+                  { field: 'agent.name', label: 'Agent name' },
+                  { field: 'rule.description', label: 'Description' },
+                  { field: 'rule.level', label: 'Level' },
+                  { field: 'rule.id', label: 'Rule ID' },
+                ]}
+                initialAgentColumns={[
+                  { field: 'icon' },
+                  { field: 'timestamp' },
+                  { field: 'rule.description', label: 'Description' },
+                  { field: 'rule.level', label: 'Level' },
+                  { field: 'rule.id', label: 'Rule ID' },
                 ]}
                 includeFilters="vulnerability"
                 implicitFilters={implicitFilters}