+++ public/components/common/modules/panel/components/agg-table.tsx
@@ -11,9 +11,9 @@
  * Find more information about this on the LICENSE file.
  */
 
-import { EuiBasicTable, EuiPanel, EuiTitle, EuiBasicTableColumn } from '@elastic/eui';
+import { EuiPanel, EuiTitle, EuiBasicTableColumn, EuiInMemoryTable } from '@elastic/eui';
 import { useEsSearch } from '../../../hooks';
-import React from 'react';
+import React, { useState, useMemo } from 'react';
 
 export const AggTable = ({
   onRowClick = (field, value) => {},
@@ -24,27 +24,33 @@ export const AggTable = ({
   panelProps,
   titleProps,
 }) => {
-  const preAppliedAggs = {
-    buckets: {
-      terms: {
-        field: aggTerm,
-        size: maxRows,
-        order: { _count: 'desc' },
+  const [order, setOrder] = useState({ _count: 'desc' });
+  const preAppliedAggs = useMemo(() => {
+    return {
+      buckets: {
+        terms: {
+          field: aggTerm,
+          size: maxRows,
+          order,
+        },
       },
-    },
-  };
+    };
+  }, [order, aggTerm, maxRows]);
+
   const { esResults, isLoading, error } = useEsSearch({ preAppliedAggs });
   const buckets = ((esResults.aggregations || {}).buckets || {}).buckets || [];
   const columns: EuiBasicTableColumn<any>[] = [
     {
       field: 'key',
       name: aggLabel,
+      sortable: true,
     },
     {
       field: 'doc_count',
       name: 'Count',
       isExpander: false,
       align: 'right',
+      sortable: true,
     },
   ];
   const getRowProps = (item) => {
@@ -56,18 +62,36 @@ export const AggTable = ({
       },
     };
   };
-
+  const pagination = {
+    hidePerPageOptions: true,
+    pageSize: 10,
+  };
+  const sorting = {
+    sort: {
+      field: 'doc_count',
+      direction: 'desc',
+    },
+  };
+  const onTableChange = ({ sort = {} }) => {
+    if (sort.field) {
+      const field = { key: '_key', doc_count: '_count' }[sort.field];
+      setOrder({ [field]: sort.direction });
+    }
+  };
   return (
     <EuiPanel data-test-subj={`${aggTerm}-aggTable`} {...panelProps}>
       <EuiTitle {...titleProps}>
         <h2>{tableTitle}</h2>
       </EuiTitle>
-      <EuiBasicTable
-        items={buckets}
+      <EuiInMemoryTable
         columns={columns}
-        rowProps={getRowProps}
+        items={buckets}
         loading={isLoading}
+        rowProps={getRowProps}
         error={error ? error.message : undefined}
+        pagination={pagination}
+        onTableChange={onTableChange}
+        sorting={sorting}
       />
     </EuiPanel>
   );